+++ sources/xcode/App/App/_build/app.js
@@ -1607,8 +1607,7 @@
 
   // JS/config/dev.js
   var Config = {
-    env: "dev",
-    allowed: ["index.html", "google.cl"]
+    env: "dev"
   };
   var dev_default = Config;
 
@@ -1802,7 +1801,7 @@
   var MainScreen = class extends component_default {
     constructor() {
       super(...arguments);
-      __publicField(this, "url", "https://google.cl");
+      __publicField(this, "url", "res://index.html");
       __publicField(this, "components", {
         pull: {
           title: i18n_default("main.pull.title", "Pull to Refresh"),