+++ sources/xcode/App/JS/config/dev.js
@@ -1,6 +1,9 @@
 const Config = {
     env: "dev",
-    allowed: ["index.html", "google.cl"],
+    // List the allowed urls.
+    // Otherwise it will launch native browser
+    // If not present will allow all urls
+    allowed: ["file://", "google.cl"],
 };
 
 export default Config;