+++ sources/xcode/App/JS/config/prod.js
@@ -1,5 +1,6 @@
 const Config = {
     env: "production",
+    // allowed: ["file://", "google.cl"],
 };
 
 export default Config;