+++ sources/xcode/CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - `$contacts` extension: `$contacts.all` for usage within the WebView.
 
+- Ability to have `allowed` list of urls in configuration. Not allowed urls will launch native browser.
+
 ### Fixed
 
 - `WKWebView` triggered `appdidLoad` event more than once. Now it only triggers the event when loaded.